feat: setup jest
