+++ .gitignore
@@ -2,6 +2,8 @@
 node_modules
 storybook-static
 jsdocs-static
+__snapshots__
+coverage
 
 dist/ 
 # local env files
