+++ package.json
@@ -12,7 +12,10 @@
     "build:components": "vue-cli-service build --target lib --name front-mfe-components src/main.js",
     "lint": "vue-cli-service lint",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "test:unit": "jest --updateSnapshot",
+    "test:coverage": "jest --updateSnapshot --coverage"
+
   },
   "files": [
     "dist/*",
@@ -21,6 +24,7 @@
   "dependencies": {
     "@socheatsok78/storybook-addon-vuetify": "^0.1.8",
     "@vue/cli-service-global": "^4.5.13",
+    "@vue/test-utils": "^1.3.0",
     "core-js": "^3.6.5",
     "sass": "~1.32.0",
     "sass-loader": "^10.2.0",
@@ -35,12 +39,13 @@
     "@storybook/vue": "^6.3.6",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-unit-jest": "^4.5.15",
     "@vue/cli-service": "~4.5.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
-    "jest": "^27.2.5",
+    "jest": "^25.5.4",
     "jsdoc": "^3.6.7",
     "vue-template-compiler": "^2.6.11",
     "webpack-cli": "^4.9.0"
