+++ tests/unit/index.js
@@ -1,0 +1,4 @@
+import Vue from 'vue';
+import Vuetify from 'vuetify';
+
+Vue.use(Vuetify);
